+++ templates/modules/navigation.php
@@ -1,3 +1,0 @@
-<nav role="navigation">
-	<?php wp_nav_menu( array( 'theme_location' => 'primary', 'menu_class' => 'menu horizontal', 'container' => false, 'fallback_cb' => 'asc_default_menu' ) ); ?>
-</nav>